+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Renamed NewSvg to NewSVG
+
 ## [0.12.0] 2020-09-09
 ### Changed